reviews 1

+++ clang/runtime/dpct-rt/include/dpct/dpl_extras/dpcpp_extensions.h
@@ -568,13 +568,11 @@ private:
 /// Helper for Block Load
 template <int GROUP_THREADS, typename InputT, int ITEMS_PER_THREAD,
           typename InputIteratorT>
-__dpct_inline__ void load_blocked(int linear_tid, InputIteratorT block_itr,
+__dpct_inline__ void load_blocked(size_t linear_tid, InputIteratorT block_itr,
                                   InputT (&items)[ITEMS_PER_THREAD]) {
 #pragma unroll
   for (int ITEM = 0; ITEM < ITEMS_PER_THREAD; ITEM++) {
-
     if ((linear_tid * ITEMS_PER_THREAD) + ITEM < GROUP_THREADS) {
-
       items[ITEM] = block_itr[(linear_tid * ITEMS_PER_THREAD) + ITEM];
     }
   }
@@ -582,21 +580,19 @@ __dpct_inline__ void load_blocked(int linear_tid, InputIteratorT block_itr,
 
 template <int GROUP_THREADS, typename InputT, int ITEMS_PER_THREAD,
           typename InputIteratorT>
-__dpct_inline__ void load_striped(int linear_tid, InputIteratorT block_itr,
+__dpct_inline__ void load_striped(size_t linear_tid, InputIteratorT block_itr,
                                   InputT (&items)[ITEMS_PER_THREAD]) {
 #pragma unroll
   for (int ITEM = 0; ITEM < ITEMS_PER_THREAD; ITEM++) {
-
     if((linear_tid + (ITEM * ITEMS_PER_THREAD) < GROUP_THREADS){
       items[ITEM] = block_itr[linear_tid + (ITEM * ITEMS_PER_THREAD)];
-
     }
   }
 }
 
 template <int GROUP_THREADS, typename InputT, int ITEMS_PER_THREAD,
           typename InputIteratorT>
-__dpct_inline__ void load_warp_striped(int linear_tid, InputIteratorT block_itr,
+__dpct_inline__ void load_warp_striped(size_t linear_tid, InputIteratorT block_itr,
                                        InputT (&items)[ITEMS_PER_THREAD]) {
 
   int tid = linear_tid & 1;
@@ -604,7 +600,6 @@ __dpct_inline__ void load_warp_striped(int linear_tid, InputIteratorT block_itr,
   int warp_offset = wid * ITEMS_PER_THREAD;
 #pragma unroll
   for (int ITEM = 0; ITEM < ITEMS_PER_THREAD; ITEM++) {
-
     new (&items[ITEM]) InputT(block_itr[warp_offset + tid + (ITEM * 1)]);
   }
 }
